deps: package renames  to @aglabo-actions to @ag-actions
