+++ package.json
@@ -41,12 +41,12 @@
     "format:dprint": "dprint fmt",
     "check:dprint": "dprint check",
     "check:types": "pnpm -r run check:types",
-    "check:spells": "pnpm exec cspell --cache --cache-location .cache/cspellcache --config .vscode/cspell.json ",
+    "check:spells": "pnpm exec cspell  --config .vscode/cspell.json --cache --cache-location .cache/cspell/cSpellCache",
     "lint": "pnpm -r run lint",
     "lint:types": "pnpm -r run lint:types",
     "lint:fix": "echo 'skip where project root'",
     "lint-all": "pnpm exec eslint --config ./configs/eslint.config.js --cache --cache-location .cache/eslint-cache/esLintCache",
-    "lint-all-types": "pnpm exec eslint --config ./configs/eslint.config.typed.js ",
+    "lint-all-types": "pnpm exec eslint --config ./configs/eslint.config.typed.js  --cache --cache-location .cache/eslint-cache/esLintCache",
     "lint:secrets": "secretlint --secretlintrc ./configs/secretlint.config.yaml --secretlintignore .gitignore --maskSecrets **/*"
   }
 }
