+++ packages/@ag-actions/tool-installer/configs/commitlint.config.ts
