+++ packages/@ag-actions/tool-installer/configs/secretlint.config.yaml
