+++ packages/@ag-actions/tool-installer/configs/tsup.config.module.ts
