+++ packages/@ag-actions/tool-installer/configs/vitest.config.ci.ts
