+++ packages/@ag-actions/tool-installer/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@aglabo-actions/tool-installer",
+  "name": "@ag-actions/tool-installer",
   "private": true,
   "version": "0.1.0",
   "description": "Universal tool installer framework for GitHub Actions (eget-based).",
@@ -27,10 +27,10 @@
     "clean": "rimraf lib module .cache",
     "format:dprint": "dprint fmt",
     "check:dprint": "dprint check",
-    "check:types": "tsc --noEmit",
-    "check:spells": "pnpm exec cspell --config .vscode/cspell.json --cache --cache-location .cache/cspellcache ",
-    "lint": "pnpm exec eslint --config ./configs/eslint.config.js --cache --cache-location .cache/eslint-cache/ ",
-    "lint:types": "pnpm exec eslint --config ./configs/eslint.config.typed.js --cache --cache-location .cache/eslint-cache/ ",
+    "check:types": "tsc --noEmit --incremental",
+    "check:spells": "pnpm exec cspell --config .vscode/cspell.json --cache --cache-location ../../../.cache/cspell/cSpellCache",
+    "lint": "pnpm exec eslint --config ./configs/eslint.config.js --cache --cache-location ../../../.cache/eslint-cache/eslintCache ",
+    "lint:types": "pnpm exec eslint --config ./configs/eslint.config.typed.js --cache --cache-location ../../../.cache/eslint-cache/eslintCache",
     "lint:all": "pnpm run lint && pnpm run lint:types",
     "lint:fix": "pnpm run lint -- --fix",
     "lint:secrets": "secretlint --secretlintrc ./configs/secretlint.config.yaml --secretlintignore .gitignore --maskSecrets **/*",
