+++ packages/@ag-actions/tool-installer/src/helper/configs/__tests__/convertConfig.spec.ts
