+++ packages/@ag-actions/tool-installer/src/helper/configs/configSearchDirs.ts
