+++ packages/@ag-actions/tool-installer/src/helper/configs/convertConfig.ts
