+++ packages/@ag-actions/tool-installer/src/helper/configs/parseConfig.ts
