+++ packages/@ag-actions/tool-installer/src/utils/logger/ConsoleLogger.class.ts
