+++ packages/@ag-actions/tool-installer/src/utils/prepareInstallDirectory.ts
