+++ packages/@ag-actions/tool-installer/tsconfig.json
@@ -29,6 +29,6 @@
     // types
     // "typesRoots": [    ],
     // cache
-    "tsBuildInfoFile": ".cache/tsbuildinfo"
+    "tsBuildInfoFile": ".cache/tsbuild-cache/tsbuildinfo"
   }
 }
