--- packages/@aglabo-actions/tool-installer/.vscode/settings.json
@@ -1,15 +1,0 @@
-{
-  "eslint.useFlatConfig": true,
-  "eslint.run": "onSave",
-  // point it at your flat config file under configs/
-  "eslint.options": {
-    "overrideConfigFile": "./configs/eslint.config.js"
-  },
-  // make sure ESLint is actually enabled for TS/JS files
-  "eslint.validate": [
-    "javascript",
-    "javascriptreact",
-    "typescript",
-    "typescriptreact"
-  ]
-}
